requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
flower==2.0.1
frozenlist==1.4.0
geographiclib==2.0
geopy==2.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
humanize==4.8.0
idna==3.4
inflection==0.5.1
kombu==5.3.2
magic-filter==1.0.12
multidict==6.0.4
numpy==1.26.2
openai==0.23.1
opencage==2.3.0
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
pandas-stubs==2.1.1.230928
Pillow==10.1.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-telegram-bot==20.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.0
sqlparse==0.4.4
tornado==6.3.3
tqdm==4.66.1
types-pytz==2023.3.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
yarl==1.9.2
